Move hasRequire to has folder
