+++ src/transforms/v2-to-v3/utils/has/hasRequire.ts
