+++ src/transforms/v2-to-v3/utils/has/index.ts
@@ -1,0 +1,1 @@
+export * from "./hasRequire";
