+++ src/transforms/v2-to-v3/utils/remove/removeV2ClientModule.ts
@@ -1,7 +1,7 @@
 import { Collection, JSCodeshift } from "jscodeshift";
 
 import { getV2ClientTypeNames, getV2ServiceModulePath } from "../get";
-import { hasRequire } from "../hasRequire";
+import { hasRequire } from "../has";
 import { isV2ClientInputOutputType } from "../isV2ClientInputOutputType";
 import { removeImportIdentifierName } from "./removeImportIdentifierName";
 import { removeRequireIdentifierName } from "./removeRequireIdentifierName";
