+++ src/transforms/v2-to-v3/utils/remove/removeV2GlobalModule.ts
@@ -1,7 +1,7 @@
 import { Collection, JSCodeshift } from "jscodeshift";
 
 import { PACKAGE_NAME } from "../config";
-import { hasRequire } from "../hasRequire";
+import { hasRequire } from "../has";
 import { removeImportIdentifierName } from "./removeImportIdentifierName";
 import { removeRequireIdentifierName } from "./removeRequireIdentifierName";
